Fix regex whitespace handling in object pattern
Adjusting the regex in object_pattern_big. This fixes edge cases where we were populating "None" as the object type due to missing whitespace characters after the "obj" marker. This will change previously generated object hash values, but in my testing with ~2500 files it only changed hashes that previously had the "None" catch-all listed as the object type. 

Also printing out the offset of the current object when ftrace is set for parse_pdf_object().